Add indexer tests for receipts in a non-canonical block and refactor
